advert list adapter -> PR #15 fixes
