+++ app/src/main/java/es/architectcoders/mascotas/ui/advertlist/viewmodel/AdvertListViewModel.kt
@@ -1,18 +1,15 @@
 package es.architectcoders.mascotas.ui.advertlist.viewmodel
 
-import android.util.Log
 import androidx.lifecycle.LiveData
 import androidx.lifecycle.MutableLiveData
 import androidx.lifecycle.ViewModel
+import androidx.lifecycle.viewModelScope
 import es.architectcoders.mascotas.model.AdvertRepository
 import es.architectcoders.mascotas.model.MyFirebaseAdvert
-import es.architectcoders.mascotas.ui.common.Scope
 import kotlinx.coroutines.launch
 
-class AdvertListViewModel(private val repository: AdvertRepository) : ViewModel(),
-    Scope by Scope.Impl(){
+class AdvertListViewModel(private val repository: AdvertRepository) : ViewModel() {
 
-    private val TAG = "AdvertListViewModel"
     private val _model = MutableLiveData<UiModel>()
     val model: LiveData<UiModel>
         get() {
@@ -27,30 +24,19 @@ class AdvertListViewModel(private val repository: AdvertRepository) : ViewModel(
         class Error(val errorString: String) : UiModel()
     }
 
-    init {
-        initScope()
-    }
-
     private fun refresh() {
-        launch {
+        viewModelScope.launch {
             _model.value = UiModel.Loading
             _model.value = UiModel.Content(repository.findRelevantAdverts())
         }
     }
 
-    override fun onCleared() {
-        destroyScope()
-        super.onCleared()
-    }
-
     fun onAdvertClicked(advert: MyFirebaseAdvert) {
-        Log.d(TAG, "onAdvertClicked ${advert.id}")
-//        _model.value = UiModel.Navigation(advert)
+        TODO("Open advert ${advert.id}")
     }
 
     fun onAdvertFavClicked(advert: MyFirebaseAdvert) {
-        Log.d(TAG, "onAdvertFavClicked ${advert.id}")
-
+        TODO("Fav advert ${advert.id}")
     }
 
 }
