+++ app/src/main/java/es/architectcoders/mascotas/ui/common/ContextExtension.kt
@@ -1,0 +1,12 @@
+package es.architectcoders.mascotas.ui.common
+
+import android.app.Activity
+import android.content.Context
+import android.content.Intent
+
+inline fun <reified T : Activity> Context.intentFor(body: Intent.() -> Unit): Intent =
+    Intent(this, T::class.java).apply(body)
+
+inline fun <reified T : Activity> Context.startActivity(body: Intent.() -> Unit) {
+    startActivity(intentFor<T>(body))
+}
