+++ app/src/main/java/es/architectcoders/mascotas/ui/common/Extensions.kt
@@ -1,72 +1,10 @@
 package es.architectcoders.mascotas.ui.common
 
-import android.app.Activity
-import android.content.Context
-import android.content.Intent
-import android.view.LayoutInflater
-import android.view.View
-import android.view.ViewGroup
-import android.widget.ImageView
-import androidx.annotation.LayoutRes
-import androidx.fragment.app.Fragment
-import androidx.lifecycle.*
-import androidx.recyclerview.widget.DiffUtil
-import androidx.recyclerview.widget.RecyclerView
-import com.bumptech.glide.Glide
-import kotlin.properties.Delegates
-
-fun ViewGroup.inflate(@LayoutRes layoutRes: Int, attachToRoot: Boolean = true): View =
-    LayoutInflater.from(context).inflate(layoutRes, this, attachToRoot)
-
-fun ImageView.loadUrl(url: String) {
-    Glide.with(context).load(url).into(this)
-}
-
-inline fun <VH : RecyclerView.ViewHolder, T> RecyclerView.Adapter<VH>.basicDiffUtil(
-    initialValue: List<T>,
-    crossinline areItemsTheSame: (T, T) -> Boolean = { old, new -> old == new },
-    crossinline areContentsTheSame: (T, T) -> Boolean = { old, new -> old == new }
-) =
-    Delegates.observable(initialValue) { _, old, new ->
-        DiffUtil.calculateDiff(object : DiffUtil.Callback() {
-            override fun areItemsTheSame(oldItemPosition: Int, newItemPosition: Int): Boolean =
-                areItemsTheSame(old[oldItemPosition], new[newItemPosition])
-
-            override fun areContentsTheSame(oldItemPosition: Int, newItemPosition: Int): Boolean =
-                areContentsTheSame(old[oldItemPosition], new[newItemPosition])
-
-            override fun getOldListSize(): Int = old.size
-
-            override fun getNewListSize(): Int = new.size
-        }).dispatchUpdatesTo(this@basicDiffUtil)
-    }
-
-inline fun <reified T : Activity> Context.intentFor(body: Intent.() -> Unit): Intent =
-    Intent(this, T::class.java).apply(body)
-
-inline fun <reified T : Activity> Context.startActivity(body: Intent.() -> Unit) {
-    startActivity(intentFor<T>(body))
-}
+import androidx.lifecycle.LifecycleOwner
+import androidx.lifecycle.LiveData
+import androidx.lifecycle.Observer
 
 fun <T, L : LiveData<T>> LifecycleOwner.observe(liveData: L, body: (T) -> Unit) {
     liveData.observe(this, Observer(body))
 }
 
-inline fun <reified T : ViewModel> Fragment.withViewModel(
-    crossinline factory: () -> T,
-    body: T.() -> Unit
-): T {
-    val vm = getViewModel(factory)
-    vm.body()
-    return vm
-}
-
-@Suppress("UNCHECKED_CAST")
-inline fun <reified T : ViewModel> Fragment.getViewModel(crossinline factory: () -> T): T {
-
-    val vmFactory = object : ViewModelProvider.Factory {
-        override fun <U : ViewModel> create(modelClass: Class<U>): U = factory() as U
-    }
-
-    return ViewModelProviders.of(this, vmFactory)[T::class.java]
-}
