+++ app/src/main/java/es/architectcoders/mascotas/ui/common/FragmentExtension.kt
@@ -1,0 +1,25 @@
+package es.architectcoders.mascotas.ui.common
+
+import androidx.fragment.app.Fragment
+import androidx.lifecycle.ViewModel
+import androidx.lifecycle.ViewModelProvider
+import androidx.lifecycle.ViewModelProviders
+
+inline fun <reified T : ViewModel> Fragment.withViewModel(
+    crossinline factory: () -> T,
+    body: T.() -> Unit
+): T {
+    val vm = getViewModel(factory)
+    vm.body()
+    return vm
+}
+
+@Suppress("UNCHECKED_CAST")
+inline fun <reified T : ViewModel> Fragment.getViewModel(crossinline factory: () -> T): T {
+
+    val vmFactory = object : ViewModelProvider.Factory {
+        override fun <U : ViewModel> create(modelClass: Class<U>): U = factory() as U
+    }
+
+    return ViewModelProviders.of(this, vmFactory)[T::class.java]
+}
