--- app/src/main/java/es/architectcoders/mascotas/ui/common/Scope.kt
@@ -1,26 +1,0 @@
-package es.architectcoders.mascotas.ui.common
-
-import kotlinx.coroutines.CoroutineScope
-import kotlinx.coroutines.Dispatchers
-import kotlinx.coroutines.Job
-import kotlinx.coroutines.SupervisorJob
-import kotlin.coroutines.CoroutineContext
-
-interface Scope : CoroutineScope {
-
-    class Impl : Scope {
-        override lateinit var job: Job
-    }
-
-    var job: Job
-    override val coroutineContext: CoroutineContext
-        get() = Dispatchers.Main + job
-
-    fun initScope() {
-        job = SupervisorJob()
-    }
-
-    fun destroyScope() {
-        job.cancel()
-    }
-}
