+++ app/src/main/java/es/architectcoders/mascotas/ui/common/ViewExtension.kt
@@ -1,0 +1,15 @@
+package es.architectcoders.mascotas.ui.common
+
+import android.view.LayoutInflater
+import android.view.View
+import android.view.ViewGroup
+import android.widget.ImageView
+import androidx.annotation.LayoutRes
+import com.bumptech.glide.Glide
+
+fun ViewGroup.inflate(@LayoutRes layoutRes: Int, attachToRoot: Boolean = true): View =
+    LayoutInflater.from(context).inflate(layoutRes, this, attachToRoot)
+
+fun ImageView.loadUrl(url: String) {
+    Glide.with(context).load(url).into(this)
+}
