+++ app/src/main/res/values/strings.xml
@@ -1,6 +1,7 @@
 <resources>
     <string name="app_name">+cotas</string>
-    <string name="advertRecentCD">Recent advert</string>
-    <string name="addAdvertCD">Add advert</string>
-    <string name="advertFavCD">Advert fav</string>
+    <string name="advert_recent">Recent advert</string>
+    <string name="add_advert">Add advert</string>
+    <string name="advert_fav">Advert fav</string>
+    <string name="advert_cover">Advert cover</string>
 </resources>
